Rewrite mkjigsnap in (almost) pure perl for better performance
Originally mkjigsnap was in shell, but performance sucked. Since the
merge of the perl generate_snapshot_tree() code, things have been
faster but the script was a mess of a combination of shell, sed, perl
etc.

Now moved over to a single perl script which (as well as being
cleaner) is measurably faster (~20-30% in my tests). Still calls out
to other programs for best performance (zcat rather than gzreadline
when reading jigdo files).

Now has an extra dependency on Compress::Zlib, aka libio-compress-perl
in Debian systems.